refactor: useChallenge 훅 분리 및 챌린지 fetch/update 기능 별도로 분리
