+++ src/app/challenge/page.tsx
@@ -2,76 +2,45 @@
 import ChallengeHeader from '@/components/challenge/ChallengeHeader';
 import ChallengeList from '@/components/challenge/ChallengeList';
 import BottomSheet from '@/components/ui/bottomsheet/BottomSheet';
-import {
-  fetchChallenges,
-  fetchCurrentChallenge,
-  resetDailyChallenge,
-  updateCurrentChallenge,
-} from '@/services/challenge/getChallenges';
 import useBottomSheetStore from '@/stores/useBottomSheetStore';
-
-import { Challenge, CurrentChallenge } from '@/types/challenge.types';
-import { isSameDate } from '@/utils/date';
-import { useEffect, useState } from 'react';
+import { Challenge } from '@/types/challenge.types';
+import { useState } from 'react';
 import ChallengeDetail from '@/components/challenge/ChallengeDetail';
+import useChallenge from '@/hooks/queries/useChallenge';
+import { useQueryClient } from '@tanstack/react-query';
+import { QUERY_KEYS } from '@/hooks/queries/queryKeys';
+import { updateCurrentChallenge } from '@/services/challenge/updateChallenges';
 
 const ChallengePage = () => {
   const { close } = useBottomSheetStore();
-
-  const [challenges, setChallenges] = useState<Challenge[] | null>(null);
   const [selectedChallenge, setSelectedChallenge] = useState<Challenge | null>(null);
-  const [todoChallenge, setTodoChallenge] = useState<Challenge | null>(null);
-
-  useEffect(() => {
-    const loadChallenges = async () => {
-      const [challenges, currentChallengeData] = await Promise.all([
-        fetchChallenges(),
-        fetchCurrentChallenge() as Promise<CurrentChallenge | null>,
-      ]);
-
-      setChallenges(challenges);
-
-      if (!currentChallengeData || !currentChallengeData.challenge_updated_at) {
-        setTodoChallenge(null);
-        return;
-      }
-
-      const updatedDate = new Date(currentChallengeData?.challenge_updated_at);
-      const now = new Date();
-
-      if (isSameDate(updatedDate, now)) {
-        setTodoChallenge(currentChallengeData.current_challenge_id ?? null);
-      } else {
-        setTodoChallenge(null);
-        await resetDailyChallenge();
-      }
-    };
-
-    loadChallenges();
-  }, []);
+  const { data, isLoading } = useChallenge();
+  const queryClient = useQueryClient();
 
   const onSelectChallenge = (id: number) => {
-    const selectChallenge = challenges?.find((challenge) => challenge.id === id) ?? null;
+    const selectChallenge = data?.challenges?.find((challenge) => challenge.id === id) ?? null;
     setSelectedChallenge(selectChallenge);
   };
 
   const handleConfirmChallenge = async () => {
     if (!selectedChallenge) return;
-    setTodoChallenge(selectedChallenge);
     await updateCurrentChallenge(selectedChallenge.id);
+    queryClient.invalidateQueries({ queryKey: QUERY_KEYS.challenge() });
     close();
   };
 
+  if (isLoading || !data) return '로딩 중입니다...';
+
   return (
     <>
-      <ChallengeHeader todoChallenge={todoChallenge} />
+      <ChallengeHeader todoChallenge={data.todoChallenge} />
       <ChallengeList
-        challenges={challenges}
+        challenges={data?.challenges}
         onSelectChallenge={onSelectChallenge}
       />
       <BottomSheet
         onClick={handleConfirmChallenge}
-        disabled={!!todoChallenge}
+        disabled={!!data?.todoChallenge}
         type='confirm'
         label={`챌린지 등록`}
       >
