+++ src/hooks/queries/queryKeys.ts
@@ -4,4 +4,5 @@ export const QUERY_KEYS = {
   authStatus: () => ['authStatus'],
   weather: (latitude: number, longitude: number) => ['weather', convertLatLngToGrid(latitude, longitude)],
   ranking: () => ['ranking'],
+  challenge: () => ['challenge'],
 };
